+++ feature/settings/src/main/kotlin/dev/hossain/devicecatalog/feature/settings/DeveloperSettingsScreenCircuit.kt
@@ -1,4 +1,4 @@
-package dev.hossain.devicecatalog.feature.statistics
+package dev.hossain.devicecatalog.feature.settings
 
 import com.slack.circuit.runtime.CircuitUiEvent
 import com.slack.circuit.runtime.CircuitUiState
@@ -6,15 +6,18 @@ import com.slack.circuit.runtime.screen.Screen
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data object AboutScreen : Screen {
+data object DeveloperSettingsScreenCircuit : Screen {
     data class State(
-        val appVersion: String,
+        val featureFlags: Map<String, Boolean>,
         val eventSink: (Event) -> Unit,
     ) : CircuitUiState
 
     sealed class Event : CircuitUiEvent {
-        data object OpenSourceInfo : Event()
+        data class ToggleFeatureFlag(
+            val key: String,
+            val value: Boolean,
+        ) : Event()
 
-        data object OpenDeveloperSettings : Event()
+        data object NavigateBack : Event()
     }
 }